https://online.sbis.ru/opendoc.html?guid=0d7c8560-60d6-413e-ad0d-fbf76055dc50 Вернул использование опции IList.contextMenuConfig
