+++ Controls/_itemActions/interface/IItemActions.ts
@@ -3,6 +3,7 @@ import {Model} from 'Types/entity';
 import {SyntheticEvent} from 'Vdom/Vdom';
 import {IBaseCollection, IItemActionsTemplateConfig, ISwipeConfig, ANIMATION_STATE} from 'Controls/display';
 import {ISource} from 'Controls/interface';
+import {IContextMenuConfig} from "./IContextMenuConfig";
 
 export enum TItemActionShowType {
     // show only in Menu
@@ -171,7 +172,7 @@ export interface IItemActionsCollection extends IBaseCollection<IItemActionsItem
     getActiveItem(): IItemActionsItem;
 }
 
-export interface IMenuTemplateOptions {
+export interface IMenuTemplateOptions extends IContextMenuConfig {
     source: ISource;
     keyProperty: string;
     parentProperty: string;
