+++ Controls/itemActions.ts
@@ -25,5 +25,6 @@ export {
     IItemAction,
     TItemActionShowType
 } from './_itemActions/interface/IItemActions';
+export {IContextMenuConfig} from './_itemActions/interface/IContextMenuConfig';
 export {Controller} from './_itemActions/Controller';
 export {Utils} from './_itemActions/Utils';
